Cleanup shouldn't remove anything

+++ unmount.sh
@@ -7,4 +7,7 @@ msg "Unmounting disk image"
 sudo umount /dev/mapper/loop0p1
 sudo kpartx -d /dev/loop0
 sudo losetup -d /dev/loop0
-rm -rf "$1"
+
+if [ $# -eq 1 ] ; then
+	rm -rf "$1"
+fi
